requirements.txt
apiview==1.3.25
asgiref==3.9.1
attrs==25.3.0
contourpy==1.3.2
cycler==0.12.1
diff-match-patch==20241021
Django==5.2.4
django-import-export==4.3.9
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-yasg==1.21.10
fonttools==4.59.0
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
matplotlib==3.10.5
numpy==2.2.6
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
response==0.5.0
rpds-py==0.26.0
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.14.1
uritemplate==4.2.0
